Removal of Unit Productions
The current algorithm is applying it rules for the fisrt production too
(S' -> S). This prodution is added to the grammar's productions as the first
step of the conversion of a CFG to its Chomsky Normal Form. For now, I don't
know if it this is correct to see this production as a Unit Production.
If not, we just have to iterate over self.productions - "root
production", instead of just self.productions.